% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev_fit}
\alias{gev_fit}
\title{Fit a GEV model}
\usage{
gev_fit(
  formulas,
  data,
  response,
  rl_n = 100,
  links = list("identity", "exponential", "identity")
)
}
\arguments{
\item{formulas}{a list with three elements, all objects of class "\code{\link[stats]{formula}}",
for the location, scale and shape parameter in that order. A symbolic description of the model}

\item{data}{data frame containing all variables listed in \code{formulas}}

\item{response}{the observed maxima corresponding to those variables in \code{data}}

\item{rl_n}{the value of n of which to calculate the return level of period n, if \code{NULL}, the return
level is not calculated}

\item{links}{a list with three elements, all strings detailing which link function to use for the
location, scale and shape parameter in that order}
}
\value{
an S3 object of class "\code{gev}", containing
\item{\code{loglik}}{value of the final log-likelihood at the values of \code{par}}
\item{\code{par}}{list of maximum likelihood parameter estimates for each GEV distribution parameter}
\item{\code{par_se}}{standard errors of the parameters in \code{par}}
\item{\code{modX}}{list of model matrices used for each parameter}
\item{\code{link}}{list of link functions inputted}
\item{\code{formula}}{list of formulae inputted}
\item{\code{big_formula}}{combination of unique elements in all formulae used}
\item{\code{data}}{data frame inputted}
\item{\code{response}}{response vector (observed maxima) inputted}
}
\description{
\code{gev_fit} is used to fit a Generalised Extreme Value (GEV) linear model. This is
a wrapper function, using \code{\link{gev_likfit}} and \code{\link{gev_grad}} to obtain parameter
estimates via maximum likelihood.
}
\details{
The \code{formulas} object have formulas specified symbolically, of the form \code{~ predictor1 + predictor2 + ...}
where both \code{predictor1} and \code{predictor2} are variables in \code{data}.

The GEV model is fit by using \code{\link[stats]{optim}} with the BFGS method.
}
