% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev_grad}
\alias{gev_grad}
\title{GEV gradient function}
\usage{
gev_grad(
  theta,
  muX,
  sigX,
  xiX,
  z,
  mulink = "identity",
  siglink = "exponential",
  xilink = "identity"
)
}
\arguments{
\item{theta}{vector of parameters, for each of location, scale and shape (in that order)}

\item{muX}{model matrix for the location parameter}

\item{sigX}{model matrix for the scale parameter}

\item{xiX}{model matrix for the shape parameter}

\item{z}{vector of observed maxima}

\item{mulink}{string describing what link function to use for the location parameter}

\item{siglink}{string describing what link function to use for the scale parameter}

\item{xilink}{string describing what link function to use for the shape parameter}
}
\value{
vector of gradients for each parameter
}
\description{
Function to calculate gradient of parameters, accounting for non-stationarity
}
\details{
For the purposes of optimisation, this function takes the same arguments as \code{\link{gev_likfit}}
so it can be passed into the \code{\link[stats]{optim}} function.
}
