% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev_optim}
\alias{gev_optim}
\title{GEV optimisation function}
\usage{
gev_optim(
  z,
  muX,
  sigX,
  xiX,
  gr = gev_grad,
  se = TRUE,
  mulink = "identity",
  siglink = "exponential",
  xilink = "identity",
  init = NULL
)
}
\arguments{
\item{z}{vector of observed maxima}

\item{muX}{model matrix for the location parameter}

\item{sigX}{model matrix for the scale parameter}

\item{xiX}{model matrix for the shape parameter}

\item{gr}{gradient function to be used in \code{\link[stats]{optim}}, if \code{NULL}, a numerical gradient approximation is used}

\item{se}{logical; if \code{TRUE}, standard error of parameters will be output too}

\item{mulink}{string describing what link function to use for the location parameter}

\item{siglink}{string describing what link function to use for the scale parameter}

\item{xilink}{string describing what link function to use for the shape parameter}

\item{init}{initial conditions for optimisation}
}
\value{
vector of gradients for each parameter
}
\description{
Optimise a non-stationary GEV distribution
}
\details{
This function uses the \code{\link[stats]{optim}} function with the BFGS method to
optimise a parameter vector corresponding to a given set of covariates for each GEV parameter.
}
