% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gev_rl}
\alias{gev_rl}
\title{Calculate Return Values}
\usage{
gev_rl(n, gevmodel, data)
}
\arguments{
\item{n}{value for which the return value is calculated for}

\item{gevmodel}{object of class "\code{gev}", fit by \code{\link{gev_fit}}}

\item{data}{data frame to find return values for, using parameter estimates from \code{gevmodel}}
}
\value{
a new data frame with an additional column \code{return}, containing the return values for the corresponding
other entries in the data frame
}
\description{
Calculates return values for each set of covariates specified in both \code{data} and the model.
}
\details{
The \eqn{n}-level return value is defined as the largest expected return over the next \eqn{n} time period.
For example, if the maxima is measured yearly, this would be the \eqn{n}-year return value.
}
