% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_oop.R
\name{plot.gev}
\alias{plot.gev}
\title{Plot GEV model diagnostics}
\usage{
\method{plot}{gev}(x, ...)
}
\arguments{
\item{x}{object of class "\code{gev}" fit by \code{\link{gev_fit}}}

\item{...}{further arguments for plotting (will apply to both plots)}
}
\value{
printed elements
}
\description{
Plot the probability and the quantile-quantile plot for the GEV distribution, for a model
fit by \code{\link{gev_fit}}.
}
\details{
The probability plot plots the model calculated
\deqn{
G(z_i) = \exp \{ - ( 1 + \xi (z_i - \mu)/\sigma)^{-1/\xi} \}
}
against the empirical
\deqn{
H(z_i) = i / (m + 1).
}
If the model is a good fit, these should be approximately equal, and so the points plotted should lie on a
diagonal line.

The quantile-quantile plot plots
\deqn{
G^{-1} = \mu - (\sigma / \xi) (1 - (- log(i/(m+1)))^{-\xi} )
}
against the observed maxima \eqn{z_i}. Similar to the probability plot, this should also lie on the diagonal line
for a good fit, as these points should be approximately equal.
}
