% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_oop.R
\name{print.gev}
\alias{print.gev}
\title{Print GEV model}
\usage{
\method{print}{gev}(x, ...)
}
\arguments{
\item{x}{object of class "\code{gev}" fit by \code{\link{gev_fit}}}

\item{...}{further arguments}
}
\value{
printed elements
}
\description{
Print elements of the GEV model.
}
\details{
This prints the maximum likelihood estimates of parameters for each parameter in the GEV distribution,
and the value of the log-likelihood here.
}
