% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain_max_model}
\alias{rain_max_model}
\title{Model Precipitation Maxima}
\format{A data frame with 4200 rows and 7 variables:
\describe{
  \item{lon}{longitude of centre of grid cell}
  \item{year}{year of estimate}
  \item{lat}{latitude of centre of grid cell}
  \item{rain}{estimated precipitation value (mm)}
  \item{id}{id number of the model, corresponding to grid cell}
  \item{elev}{average elevation across grid cell}
  \item{dist2coast}{distance in km to the coastline from the grid point centre}
}}
\usage{
data(rain_max_obs)
}
\description{
Gridded model predicted yearly precipitation maxima from the ERA5 reanalysis from the period 1979 to 2018 in South West UK
}
\keyword{datasets}
