% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain_max_obs}
\alias{rain_max_obs}
\title{Observed Precipitation Maxima}
\format{A data frame with 4731 rows and 8 variables:
\describe{
  \item{date}{date of observation}
  \item{lon}{longitude of observation}
  \item{lat}{latitude of observation}
  \item{elev}{elevation of observation}
  \item{rain}{precipitation value (mm)}
  \item{year}{year of observation}
  \item{day}{day of month of observation}
  \item{month}{month of observation}
}}
\usage{
data(rain_max_obs)
}
\description{
Observed yearly precipitation maxima from 1988 to 2017 in South West UK
}
\keyword{datasets}
